refactor(ios): refactor uimanger's code and structure
