+++ renderer/native/ios/renderer/HippyUIManager.h
@@ -26,104 +26,78 @@
 #import "NativeRenderDefines.h"
 #import "HippyBridgeModule.h"
 #import "HippyCustomTouchHandlerProtocol.h"
-#include <memory>
-#include <unordered_map>
-#include <functional>
 
-@class NativeRenderAnimationViewParams, HippyShadowView, HippyUIManager,HippyViewManager;
-@class NativeRenderReusePool, HippyComponentMap;
-
-class VFSUriLoader;
-namespace hippy {
-inline namespace dom {
-class RenderManager;
-class DomManager;
-class DomArgument;
-class RootNode;
-class DomNode;
-struct LayoutResult;
-using CallFunctionCallback = std::function<void(std::shared_ptr<DomArgument>)>;
-}
-}
-
-namespace footstone {
-inline namespace value {
-class HippyValue;
-}
-}
-
+@class NativeRenderAnimationViewParams;
+@class HippyShadowView;
+@class HippyUIManager;
+@class HippyViewManager;
+@class NativeRenderReusePool;
+@class HippyComponentMap;
 @protocol HippyImageProviderProtocol;
 
-/**
- * The NativeRenderUIManager is the module responsible for updating the view hierarchy.
- */
+
+/// The HippyUIManager responsible for updating the view hierarchy.
 @interface HippyUIManager : NSObject <HippyInvalidating>
 
++ (instancetype)new NS_UNAVAILABLE;
 - (instancetype)init NS_UNAVAILABLE;
 
+#ifdef __cplusplus
+/// Init method
+/// - Parameter renderManager: the hippy::RenderManager
 - (instancetype)initWithRenderManager:(std::weak_ptr<hippy::RenderManager>)renderManager NS_DESIGNATED_INITIALIZER;
 
-@property(nonatomic, assign) BOOL uiCreationLazilyEnabled;
-
-@property (nonatomic, weak) HippyBridge *bridge;
 @property (nonatomic, assign) std::weak_ptr<VFSUriLoader> VFSUriLoader;
 @property (nonatomic, assign) std::weak_ptr<hippy::RenderManager> renderManager;
 @property (nonatomic, readonly) std::weak_ptr<hippy::DomManager> domManager;
+#endif
+
+@property (nonatomic, weak) HippyBridge *bridge;
 @property (nonatomic, readonly) HippyComponentMap *viewRegistry;
-
-/**
- * Gets the view associated with a hippyTag.
- */
-- (UIView *)viewForHippyTag:(NSNumber *)hippyTag
-                  onRootTag:(NSNumber *)rootTag;
+@property (nonatomic, assign) BOOL uiCreationLazilyEnabled;
 
 
-/**
- * Get the shadow view associated with a hippyTag
- */
-- (HippyShadowView *)shadowViewForHippyTag:(NSNumber *)hippyTag
-                                 onRootTag:(NSNumber *)rootTag;
+/// Gets the view associated with a hippyTag.
+/// - Parameters:
+///   - hippyTag: NSNumber
+///   - rootTag: NSNumber
+- (UIView *)viewForHippyTag:(NSNumber *)hippyTag onRootTag:(NSNumber *)rootTag;
 
-/**
- * Update the frame of a view. This might be in response to a screen rotation
- * or some other layout event outside of the Hippy-managed view hierarchy.
- */
+/// Get the shadow view associated with a hippyTag
+/// - Parameters:
+///   - hippyTag: NSNumber
+///   - rootTag: NSNumber
+- (HippyShadowView *)shadowViewForHippyTag:(NSNumber *)hippyTag onRootTag:(NSNumber *)rootTag;
+
+/// Update the frame of a view. This might be in response to a screen rotation
+/// or some other layout event outside of the Hippy-managed view hierarchy.
+/// - Parameters:
+///   - frame: new frame
+///   - view: target view
 - (void)setFrame:(CGRect)frame forRootView:(UIView *)view;
 
-/**
- * Schedule a block to be executed on the UI thread. Useful if you need to execute
- * view logic after all currently queued view updates have completed.
- */
+/// Schedule a block to be executed on the UI thread. Useful if you need to execute
+/// view logic after all currently queued view updates have completed.
+/// - Parameter block: block to be executed on main thread
 - (void)addUIBlock:(HippyViewManagerUIBlock)block;
 
-/**
- * In some cases we might want to trigger layout from native side.
- * Hippy won't be aware of this, so we need to make sure it happens.
- */
+/// In some cases we might want to trigger layout from native side.
+/// Hippy won't be aware of this, so we need to make sure it happens.
+/// - Parameter tag: root tag
 - (void)setNeedsLayoutForRootNodeTag:(NSNumber *)tag;
 
-- (void)registerRootView:(UIView *)rootView asRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-- (void)unregisterRootViewFromTag:(NSNumber *)rootTag;
-
+/// Get all rootView
 - (NSArray<__kindof UIView *> *)rootViews;
 
-- (__kindof UIView *)viewFromRenderViewTag:(NSNumber *)componentTag onRootTag:(NSNumber *)rootTag;
-
+/// Purge view from superView
 - (void)purgeViewsFromComponentTags:(NSArray<NSNumber *> *)componentTag onRootTag:(NSNumber *)rootTag;
 
+/// Update view with props
 - (void)updateView:(NSNumber *)componentTag onRootTag:(NSNumber *)rootTag props:(NSDictionary *)pros;
 
+/// Get viewManager Name
+/// @param viewName NSString
 - (__kindof HippyViewManager *)renderViewManagerForViewName:(NSString *)viewName;
-
-/**
- * Manully create views recursively from hippy tag
- *
- * @param hippyTag hippy tag corresponding to UIView
- * @return view created by hippy tag
- */
-- (UIView *)createViewRecursivelyFromcomponentTag:(NSNumber *)hippyTag
-                                        onRootTag:(NSNumber *)rootTag;
 
 /**
  * Manully create views recursively from renderObject
@@ -133,125 +107,17 @@ class HippyValue;
  */
 - (UIView *)createViewRecursivelyFromRenderObject:(HippyShadowView *)renderObject;
 
-/**
- * set dom manager for NativeRenderUIManager which holds a weak reference to domManager
- */
-- (void)setDomManager:(std::weak_ptr<hippy::DomManager>)domManager;
-
-/**
- *  create views from dom nodes
- *  @param nodes A set of nodes for creating views
- */
-- (void)createRenderNodes:(std::vector<std::shared_ptr<hippy::DomNode>> &&)nodes
-               onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- *  update views' properties from dom nodes
- *  @param nodes A set of nodes for updating views' properties
- */
-- (void)updateRenderNodes:(std::vector<std::shared_ptr<hippy::DomNode>> &&)nodes
-               onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- *  delete views from dom nodes
- *  @param nodes nodes to delete
- */
-- (void)deleteRenderNodesIds:(std::vector<std::shared_ptr<hippy::DomNode>> &&)nodes
-                  onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- * move views from container to another container
- *
- * @param ids A set of nodes id to move
- * @param fromContainer Source view container from which views move
- * @param toContainer Target view container to which views move
- */
-- (void)renderMoveViews:(const std::vector<int32_t> &&)ids
-          fromContainer:(int32_t)fromContainer
-            toContainer:(int32_t)toContainer
-                  index:(int32_t)index
-             onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-- (void)renderMoveNodes:(std::vector<std::shared_ptr<hippy::DomNode>> &&)nodes
-             onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-/**
- * update layout for view
- *
- * @param layoutInfos Vector for nodes layout infos
- */
-- (void)updateNodesLayout:(const std::vector<std::tuple<int32_t, hippy::LayoutResult>> &)layoutInfos
-               onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- * Invoked after batched operations completed
- */
-- (void)batchOnRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- * call function of view
- *
- * @param functionName Function Name to be invoked
- * @param viewName Name of target view whose function invokes
- * @param hippyTag id of target view whose function invokes
- * @param params parameters of function to be invoked
- * @param cb A callback for the return value of function
- *
- * @result Function return result
- */
-- (id)dispatchFunction:(const std::string &)functionName
-              viewName:(const std::string &)viewName
-               viewTag:(int32_t)hippyTag
-            onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode
-                params:(const footstone::value::HippyValue &)params
-              callback:(hippy::CallFunctionCallback)cb;
-
+/// Register extra components
+/// @param extraComponents extra components classes
 - (void)registerExtraComponent:(NSArray<Class> *)extraComponents;
 
-/**
- * register event for specific view
- *
- * @param name event name
- * @param node_id id for node for the event
- */
-- (void)addEventName:(const std::string &)name
-        forDomNodeId:(int32_t)node_id
-          onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- * unregister event for specific view
- *
- * @param name event name
- * @param node_id node id for the event
- */
-- (void)removeEventName:(const std::string &)name
-           forDomNodeId:(int32_t)node_id
-             onRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- * unregister vsync event
- */
-- (void)removeVSyncEventOnRootNode:(std::weak_ptr<hippy::RootNode>)rootNode;
-
-/**
- * Set root view size changed event callback
- *
- *@param cb callback
- */
-- (void)setRootViewSizeChangedEvent:(std::function<void(int32_t rootTag, NSDictionary *)>)cb;
-
-/**
- * clear all resources
- */
+/// Clear all resources
 - (void)invalidate;
-
-#if HIPPY_DEBUG
-@property(nonatomic, assign) std::unordered_map<int32_t, std::unordered_map<int32_t, std::shared_ptr<hippy::DomNode>>> domNodesMap;
-- (std::shared_ptr<hippy::DomNode>)domNodeForTag:(int32_t)dom_tag onRootNode:(int32_t)root_tag;
-- (std::vector<std::shared_ptr<hippy::DomNode>>)childrenForNodeTag:(int32_t)tag onRootNode:(int32_t)root_tag;
-#endif
 
 @end
 
+
+#pragma mark - HippyBridge (HippyUIManager)
 
 /**
  * This category makes the current HippyUIManager instance available via the
